Format the address like the widget is expecting
